updated likes table columns to int from bigint

+++ db/migrations/000001_init_schema.up.sql
@@ -21,8 +21,8 @@ CREATE TABLE "beats" (
 
 CREATE TABLE "likes" (
     "id" SERIAL PRIMARY KEY,
-    "user_id" bigint NOT NULL,
-    "beat_id" bigint NOT NULL
+    "user_id" integer NOT NULL,
+    "beat_id" integer NOT NULL
 );
 
 ALTER TABLE
